refactor: changed phone routes to reflect simpler module structure
