+++ web-client/src/modules/phone/constants.tsx
@@ -1,3 +1,5 @@
 import Location from 'react-app-location';
 
 export const PhoneEntryLocation = new Location('/phone/entry');
+
+export const PhoneVerifyLocation = new Location('/phone/verify');
