+++ web-client/src/modules/phone/index.ts
@@ -1,9 +1,9 @@
 import { Module } from 'src/types/module';
 
-import ContentPage from './pages/ContentPage';
+import Routes from './routes';
 
 const phoneModule: Module = {
   path: '/phone',
-  component: ContentPage,
+  component: Routes,
 };
 export default phoneModule;
