--- web-client/src/modules/phone/pages/routes/PhoneVerifyRoute/constants.tsx
@@ -1,3 +1,0 @@
-import Location from 'react-app-location';
-
-export const PhoneVerifyLocation = new Location('/phone/verify');
